requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
psycopg2
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
